+++ src/Daedalus.jl
@@ -12,6 +12,6 @@ include("Model.jl")
 include("Ensemble.jl")
 include("Outputs.jl")
 
-export Npi, ReactiveEffect, TimedEffect
+export Npi, ParamEffect, ReactiveTrigger, TimeTrigger, get_values, get_time
 
 end